requirements.txt
aiogram==2.21
Pillow==9.0.1
